Fixes Python missing references
* When a Python object in Lua is ressurected, it may result in a missing
reference because its reference count has been zeroed and Python can't
ressurect it like Lua. This would lead to the dangling pointer problem.
The way to solve this is by tagging the Python object in Lua when it is
deallocated so that if it is later accessed, an error is raised.
* Optimized `LuaRuntime._pyrefs_in_lua`: Created the Cython class called
`_PyReference`, which contains the Python object being referenced and
the reference count (from Lua `userdata`)
* Added tests for missing references
* Added `lua_State*` parameter to `_LuaObject.push_lua_object` in order to
call `lua_rawgeti` with the correct Lua state, and substituted calls to
`lua_rawgeti` to `_LuaObject.push_lua_object` because it checks whether the
object is missing!
* Added `unpack_python_argument_or_jump` for checking whether the Python
object is missing and doing a long jump if so. This makes the code less
repetitive too.