Change Python thread ids to "pythread_t", which is unsigned in Py3.7+ (GH-165)
Since Python 3.7, thread ids are unsigned. Before, they were signed.
So we cannot rely on storing '-1' as a sentinel value any more but use `_count > 0` as marker for a valid owner.
Also, "_owner" is not reset on release because it is only consulted when _count > 0, so that iff _count > 0, _owner is valid.
Also, make the acquisition counters unsigned.