fix unrefereing all objects created in coroutine
Unreferencing all objects created in coroutine is not good solution because such objects can be used outside of coroutine scope.
This fix refactor all reference tracking to tracking objects and when coroutine is done, set state of all objects that was created in coroutine to parent lua state. This way
objects created in coroutine can be used outside of coroutine scope and it does not cause segmentation fault.